Drop bytebuffer-compact.patch
The bytebuffer-compact patch adds a (package-private) method clearMark to reset
the mark.  Upstream already addressed this problem:
6593946: (bf) X-Buffer.compact() does not discard mark as specified.
The upstream fix adds a method discardMark(), and updates the rest of the JDK
to use that method. The clearMark() method added by this patch is not used
anywhere and can be dropped safely.